Refactor structure of cache objects into their own files
